added a launch configuration for remote provisioning of an eclipse feature
This commit adds basic support for remote provisioning using the Eclipse Launch Framework. It consists of 2 new plugins: the org.eclipse.rtp.httpdeployer.ui.launch plugin provides ui code for the launcher; the org.eclipse.rtp.httpdeployer.launch plugin is in charge of the objects needed to connect and interact with a remote httpdeployer instance.